requirements.txt
Django==4.1.7
Pillow==9.4.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
django-ckeditor==6.5.1
django-filter==22.1
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.5